+++ tasks.rb
@@ -140,6 +140,11 @@ def create_hpxmls
     'base-dhw-tank-heat-pump-uef.xml' => 'base-dhw-tank-heat-pump.xml',
     'base-dhw-tank-heat-pump-with-solar.xml' => 'base-dhw-tank-heat-pump.xml',
     'base-dhw-tank-heat-pump-with-solar-fraction.xml' => 'base-dhw-tank-heat-pump.xml',
+    'base-dhw-tank-heat-pump-operating-mode-heat-pump-only.xml' => 'base-dhw-tank-heat-pump-uef.xml',
+    'base-dhw-tank-heat-pump-detailed-schedules.xml' => 'base-dhw-tank-heat-pump-uef.xml',
+    'base-dhw-tank-model-type-stratified.xml' => 'base.xml',
+    'base-dhw-tank-detailed-setpoints.xml' => 'base.xml',
+    'base-dhw-tank-model-type-stratified-detailed-occupancy-stochastic.xml' => 'base-dhw-tank-model-type-stratified.xml',
     'base-dhw-tank-oil.xml' => 'base-dhw-tank-gas.xml',
     'base-dhw-tank-wood.xml' => 'base-dhw-tank-gas.xml',
     'base-dhw-tankless-electric.xml' => 'base.xml',
@@ -150,6 +155,7 @@ def create_hpxmls
     'base-dhw-tankless-gas-with-solar.xml' => 'base-dhw-tankless-gas.xml',
     'base-dhw-tankless-gas-with-solar-fraction.xml' => 'base-dhw-tankless-gas.xml',
     'base-dhw-tankless-propane.xml' => 'base-dhw-tankless-gas.xml',
+    'base-dhw-tankless-detailed-setpoints.xml' => 'base-dhw-tankless-gas.xml',
     'base-enclosure-2stories.xml' => 'base.xml',
     'base-enclosure-2stories-garage.xml' => 'base-enclosure-2stories.xml',
     'base-enclosure-beds-1.xml' => 'base.xml',
@@ -515,7 +521,6 @@ def set_measure_argument_values(hpxml_file, args, sch_args, orig_parent)
   args['apply_validation'] = true
 
   if ['base.xml'].include? hpxml_file
-    args['occupancy_calculation_type'] = HPXML::OccupancyCalculationTypeAsset
     args['simulation_control_timestep'] = 60
     args['site_iecc_zone'] = '5B'
     args['site_state_code'] = 'CO'
@@ -844,7 +849,6 @@ def set_measure_argument_values(hpxml_file, args, sch_args, orig_parent)
     args['hot_tub_heater_annual_therm'] = Constants.Auto
     args['hot_tub_heater_usage_multiplier'] = 1.0
   elsif ['ASHRAE_Standard_140/L100AC.xml'].include? hpxml_file
-    args['occupancy_calculation_type'] = HPXML::OccupancyCalculationTypeAsset
     args['weather_station_epw_filepath'] = 'USA_CO_Colorado.Springs-Peterson.Field.724660_TMY3.epw'
     args['geometry_unit_type'] = HPXML::ResidentialTypeSFD
     args['geometry_unit_cfa'] = 1539.0
@@ -1607,6 +1611,10 @@ def set_measure_argument_values(hpxml_file, args, sch_args, orig_parent)
     args['water_heater_type'] = HPXML::WaterHeaterTypeTankless
     args['water_heater_tank_volume'] = Constants.Auto
     args['water_heater_efficiency'] = 0.99
+  elsif ['base-dhw-tank-heat-pump-operating-mode-heat-pump-only.xml'].include? hpxml_file
+    args['water_heater_operating_mode'] = HPXML::WaterHeaterOperatingModeHeatPumpOnly
+  elsif ['base-dhw-tank-model-type-stratified.xml'].include? hpxml_file
+    args['water_heater_tank_model_type'] = HPXML::WaterHeaterTankModelTypeStratified
   end
 
   # Enclosure
@@ -2326,6 +2334,10 @@ def set_measure_argument_values(hpxml_file, args, sch_args, orig_parent)
     args['whole_house_fan_present'] = true
     args['whole_house_fan_flow_rate'] = Constants.Auto
     args['whole_house_fan_power'] = Constants.Auto
+    args['hvac_control_heating_weekday_setpoint'] = Constants.Auto
+    args['hvac_control_heating_weekend_setpoint'] = Constants.Auto
+    args['hvac_control_cooling_weekday_setpoint'] = Constants.Auto
+    args['hvac_control_cooling_weekend_setpoint'] = Constants.Auto
   elsif ['base-misc-emissions.xml'].include? hpxml_file
     args['emissions_scenario_names'] = 'Cambium Hourly MidCase LRMER RMPA, Cambium Hourly LowRECosts LRMER RMPA, Cambium Annual MidCase AER National, eGRID RMPA, eGRID RMPA'
     args['emissions_types'] = 'CO2e, CO2e, CO2e, SO2, NOx'
@@ -2504,19 +2516,57 @@ def set_measure_argument_values(hpxml_file, args, sch_args, orig_parent)
 
   # Setpoint Schedules
   if ['base-schedules-detailed-setpoints.xml'].include? hpxml_file
+    args['hvac_control_heating_weekday_setpoint'] = Constants.Auto
+    args['hvac_control_heating_weekend_setpoint'] = Constants.Auto
+    args['hvac_control_cooling_weekday_setpoint'] = Constants.Auto
+    args['hvac_control_cooling_weekend_setpoint'] = Constants.Auto
     args['schedules_filepaths'] = '../../HPXMLtoOpenStudio/resources/schedule_files/setpoints.csv'
   elsif ['base-schedules-detailed-setpoints-daily-schedules.xml'].include? hpxml_file
+    args['hvac_control_heating_weekday_setpoint'] = Constants.Auto
+    args['hvac_control_heating_weekend_setpoint'] = Constants.Auto
+    args['hvac_control_cooling_weekday_setpoint'] = Constants.Auto
+    args['hvac_control_cooling_weekend_setpoint'] = Constants.Auto
     args['schedules_filepaths'] = '../../HPXMLtoOpenStudio/resources/schedule_files/setpoints-daily-schedules.csv'
   elsif ['base-schedules-detailed-setpoints-daily-setbacks.xml'].include? hpxml_file
+    args['hvac_control_heating_weekday_setpoint'] = Constants.Auto
+    args['hvac_control_heating_weekend_setpoint'] = Constants.Auto
+    args['hvac_control_cooling_weekday_setpoint'] = Constants.Auto
+    args['hvac_control_cooling_weekend_setpoint'] = Constants.Auto
     args['schedules_filepaths'] = '../../HPXMLtoOpenStudio/resources/schedule_files/setpoints-daily-setbacks.csv'
   elsif ['base-schedules-detailed-all-10-mins.xml'].include? hpxml_file
+    args['hvac_control_heating_weekday_setpoint'] = Constants.Auto
+    args['hvac_control_heating_weekend_setpoint'] = Constants.Auto
+    args['hvac_control_cooling_weekday_setpoint'] = Constants.Auto
+    args['hvac_control_cooling_weekend_setpoint'] = Constants.Auto
     args['schedules_filepaths'] = '../../HPXMLtoOpenStudio/resources/schedule_files/setpoints-10-mins.csv'
-    sch_args['hpxml_path'] = args['hpxml_path']
-    sch_args['hpxml_output_path'] = sch_args['hpxml_path']
   elsif ['base-hvac-furnace-gas-only-detailed-setpoints.xml'].include? hpxml_file
+    args['hvac_control_heating_weekday_setpoint'] = Constants.Auto
+    args['hvac_control_heating_weekend_setpoint'] = Constants.Auto
     args['schedules_filepaths'] = '../../HPXMLtoOpenStudio/resources/schedule_files/setpoints-heating-only.csv'
   elsif ['base-hvac-room-ac-only-detailed-setpoints.xml'].include? hpxml_file
+    args['hvac_control_cooling_weekday_setpoint'] = Constants.Auto
+    args['hvac_control_cooling_weekend_setpoint'] = Constants.Auto
     args['schedules_filepaths'] = '../../HPXMLtoOpenStudio/resources/schedule_files/setpoints-cooling-only.csv'
+  end
+
+  # Water Heater Schedules
+  if ['base-dhw-tank-heat-pump-detailed-schedules.xml'].include? hpxml_file
+    args['water_heater_setpoint_temperature'] = Constants.Auto
+    args['schedules_filepaths'] = '../../HPXMLtoOpenStudio/resources/schedule_files/water-heater-setpoints.csv, ../../HPXMLtoOpenStudio/resources/schedule_files/water-heater-operating-modes.csv'
+  elsif ['base-dhw-tank-detailed-setpoints.xml'].include? hpxml_file
+    args['water_heater_setpoint_temperature'] = Constants.Auto
+    args['schedules_filepaths'] = '../../HPXMLtoOpenStudio/resources/schedule_files/water-heater-setpoints.csv'
+  elsif ['base-dhw-tankless-detailed-setpoints.xml'].include? hpxml_file
+    args['water_heater_setpoint_temperature'] = Constants.Auto
+    args['schedules_filepaths'] = '../../HPXMLtoOpenStudio/resources/schedule_files/water-heater-setpoints.csv'
+  elsif ['base-dhw-tank-model-type-stratified-detailed-occupancy-stochastic.xml'].include? hpxml_file
+    sch_args['hpxml_path'] = args['hpxml_path']
+    sch_args['schedules_type'] = 'stochastic'
+    sch_args['output_csv_path'] = '../../HPXMLtoOpenStudio/resources/schedule_files/occupancy-stochastic.csv'
+    sch_args['hpxml_output_path'] = sch_args['hpxml_path']
+  elsif ['base-schedules-detailed-all-10-mins.xml'].include? hpxml_file
+    args['water_heater_setpoint_temperature'] = Constants.Auto
+    args['schedules_filepaths'] += ', ../../HPXMLtoOpenStudio/resources/schedule_files/water-heater-setpoints-10-mins.csv'
   end
 end